refactor: extract redaction logic to iaqualink.utils.redact
Consolidates all redaction into a single module so library logging
and capture output share the same implementation. _redact_dict and
_redact_value accept an optional keys_ci param; capture passes its
extended set (adds set-cookie/state/username); library uses the base
_REDACT_KEYS_CI. Library logging now gets recursive, case-insensitive,
substring-matching, and email-masking redaction instead of the old
flat case-sensitive dict comprehension.
